requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
aliyun-python-sdk-core==2.13.16
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.10.1
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crcmod==1.7
cryptography==2.9.2
cssselect==1.1.0
fonttools==4.9.0
func-timeout==4.3.5
httplib2==0.18.1
idna==2.8
jmespath==0.9.5
Js2Py==0.70
lxml==4.5.1
multidict==4.7.4
Naked==0.1.31
oss2==2.11.0
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.7
pyjsparser==2.7.1
PyMySQL==0.9.2
pytz==2020.1
PyYAML==5.3.1
qiniu==7.2.8
redis==3.5.2
requests==2.22.0
rsa==4.6
selenium==3.141.0
shellescape==3.8.1
six==1.14.0
soupsieve==1.9.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.8
yarl==1.4.2
